rename client2 module to client
