+++ gossip4/client/client.go
@@ -1,4 +1,4 @@
-package client2
+package client
 
 import (
 	"context"
