+++ gossip4/client/roundsubscriber.go
@@ -1,4 +1,4 @@
-package client2
+package client
 
 import (
 	"context"
